Archimedean Copulas

AMHCopula

TwinCopulas.AMHCopula — Type

AMHCopula{P}

Fields:

θ::Real - parameter

Constructor

AMHCopula(θ)

The bivariate AMH copula is parameterized by $\theta \in [-1,1)$. It is an Archimedean copula with generator :

\[\phi(t) = 1 - \frac{1-\theta}{e^{-t}-\theta}\]

It has a few special cases:

When θ = 0, it is the IndependentCopula

When θ = 1, it is the UtilCopula

References:

Nelsen, Roger B. An introduction to copulas. Springer, 2006.

ClaytonCopula

TwinCopulas.ClaytonCopula — Type

ClaytonCopula{P}

Fields:

θ::Real - parameter

Constructor

ClaytonCopula(d,θ)

The bivariate Clayton copula is parameterized by $\theta \in [-1,\infty)$. It is an Archimedean copula with generator :

\[\phi(t) = \left(1+\mathrm{sign}(\theta)*t\right)^{-1\frac{1}{\theta}}\]

It has a few special cases:

When θ = -1, it is the WCopula (Lower Frechet-Hoeffding bound)

When θ = 0, it is the IndependentCopula

When θ = 1, it is the UtilCopula

When θ = ∞, is is the MCopula (Upper Frechet-Hoeffding bound)

References:

Nelsen, Roger B. An introduction to copulas. Springer, 2006.

FrankCopula

TwinCopulas.FrankCopula — Type

FrankCopula{P}

Fields:

θ::Real - parameter

Constructor

FrankCopula(θ)

The bivariate Frank copula is parameterized by $\theta \in (-\infty,\infty)$. It is an Archimedean copula with generator :

\[\phi(t) = -\frac{\log\left(1+e^{-t}(e^{-\theta-1})\right)}{ heta}\]

It has a few special cases:

When θ = -∞, it is the WCopula (Lower Frechet-Hoeffding bound)

When θ = 1, it is the IndependentCopula

When θ = ∞, is is the MCopula (Upper Frechet-Hoeffding bound)

References:

Nelsen, Roger B. An introduction to copulas. Springer, 2006.

GumbelBarnettCopula

TwinCopulas.GumbelBarnettCopula — Type

GumbelBarnettCopula{P}

Fields:

θ::Real - parameter

Constructor

GumbelBarnettCopula(θ)

The bivariate Gumbel-Barnett copula is parameterized by $\theta \in (0,1]$. It is an Archimedean copula with generator :

\[\phi(t) = \exp{θ^{-1}(1-e^{t})}, 0 \leq \theta \leq 1.\]

It has a few special cases:

When θ = 0, it is the IndependentCopula

References:

Joe, H. (2014). Dependence modeling with copulas. CRC press, Page.437

Nelsen, Roger B. An introduction to copulas. Springer, 2006.

GumbelCopula

TwinCopulas.GumbelCopula — Type

GumbelCopula{P}

Fields:

θ::Real - parameter

Constructor

GumbelCopula(d,θ)

The bivariate Gumbel copula is parameterized by $\theta \in [1,\infty)$. It is an Archimedean copula with generator :

\[\phi(t) = \exp{-t^{\frac{1}{θ}}}\]

It has a few special cases:

When θ = 1, it is the IndependentCopula

When θ = ∞, is is the MCopula (Upper Frechet-Hoeffding bound)

References:

Nelsen, Roger B. An introduction to copulas. Springer, 2006.

InvGaussianCopula

TwinCopulas.InvGaussianCopula — Type

InvGaussianCopula{P}

Fields:

θ::Real - parameter

Constructor

InvGaussianCopula(θ)

The bivariate Inverse Gaussian copula is parameterized by $\theta \in [0,\infty)$. It is an Archimedean copula with generator :

\[\phi(t) = \exp{\frac{1-\sqrt{1+2θ^{2}t}}{θ}}.\]

More details about Inverse Gaussian Archimedean copula are found in :

Mai, Jan-Frederik, and Matthias Scherer. Simulating copulas: stochastic models, sampling algorithms, and applications. Vol. 6. # N/A, 2017. Page 74.

It has a few special cases:

When θ = 0, it is the IndependentCopula

References:

Nelsen, Roger B. An introduction to copulas. Springer, 2006.

JoeCopula

TwinCopulas.JoeCopula — Type

JoeCopula{P}

Fields:

θ::Real - parameter

Constructor

JoeCopula(θ)

The bivariate Joe copula is parameterized by $\theta \in [1,\infty)$. It is an Archimedean copula with generator :

\[\phi(t) = 1 - \left(1 - e^{-t}\right)^{\frac{1}{\theta}}\]

It has a few special cases:

When θ = 1, it is the IndependentCopula

When θ = ∞, is is the MCopula (Upper Frechet-Hoeffding bound)

References:

Nelsen, Roger B. An introduction to copulas. Springer, 2006.

Nelsen2Copula

TwinCopulas.Nelsen2Copula — Type

Nelsen2Copula{P}

Fields:

θ::Real - parameter

Constructor

Nelsen2Copula(θ)

The bivariate Nelsen2Copula copula is parameterized by $\theta \in [1,\infty)$. It is an Archimedean copula with generator :

\[\phi(t) = 1 - t^{\frac{1}{\theta}}\]

It has a few special cases:

When θ = 1, it is the IndependentCopula

When θ = ∞, is is the MCopula (Upper Frechet-Hoeffding bound)

References:

Nelsen, Roger B. An introduction to copulas. Springer, 2006.

UtilCopula

TwinCopulas.UtilCopula — Type

UtilCopula{}

Constructor

UtilCopula()

The bivariate UtilCopula is a simple copula that appears as a special case of several copulas, that has the form :

\[C(u_1, u_2) = \frac{u_1u_2}{u_1+u_2 - u_1u_2}\]

It happends to be an Archimedean Copula, with generator :

\[\phi(t) = 1 / (t + 1)\]

References:

Nelsen, Roger B. An introduction to copulas. Springer, 2006.