Elliptical Copulas

GaussianCopula

TwinCopulas.GaussianCopula — Type

GaussianCopula{P}

Fields:

θ::Real - Parameter

Constructor

GaussianCopula(θ)

The bivariate Gaussian copula. It is constructed as:

\[C(u_1, u_2; \theta) = \Phi_{\theta}(\Phi^{-1}(u_1),\Phi^{-1}(u_2))\]

where $\Phi_{\theta}$ is the cumulative distribution function (CDF) of a standard bivariate normal distribution with correlation coefficient $\theta \in [-1, 1]$ and $\Phi^{-1}$is the quantile function of the standard normal distribution.

It has a few special cases:

When θ = -1, it is the WCopula (Lower Frechet-Hoeffding bound)

When θ = 0, it is the IndependentCopula

When θ = 1, is is the MCopula (Upper Frechet-Hoeffding bound)

References:

Nelsen, Roger B. An introduction to copulas. Springer, 2006.

tCopula

TwinCopulas.tCopula — Type

tCopula{P}

Fields: - ν::Real - paremeter - θ::Real - Parameter

Constructor

tCopula(ν, θ)

The bivariate t copula. It is constructed as:

\[C(u_1, u_2; \nu, \theta) = t_{\nu, \theta}(t_{\nu}^{-1}(u_1),t_{\nu}^{-1}(u_2))\]

where $t_{\nu, \theta}$ is the cumulative distribution function (CDF) of a bivariate t-distribution with $\nu \in \mathbb{R}^{+}$ degrees of freedom, zero means, and correlation $\theta \in [-1, 1]$, and $t_{\nu}^{-1}$ is the quantile function of the standard t-distribution with $\nu$ degrees of freedom.

It has a few special cases:

When θ = -1, it is the WCopula (Lower Frechet-Hoeffding bound)

When θ = 0, it is the IndependentCopula

When θ = 1, is is the MCopula (Upper Frechet-Hoeffding bound)

References:

oe, Harry. Dependence modeling with Copulas. Chapman & Hall, 2014.