Extreme Values Copulas

AsymLogCopula

TwinCopulas.AsymLogCopula — Type

AsymLogCopula{P}

Fields:

α::Real - Dependency parameter

θ::Vector - Asymmetry parameters (size 2)

Constructor

AsymLogCopula(α, θ)

The Asymmetric bivariate Logistic copula is parameterized by one dependence parameter $\alpha \in [1, \infty]$ and two asymmetry parameters $\theta_{i} \in [0,1], i=1,2$. It is an Extreme value copula with Pickands dependence function:

\[A(t) = (\theta_1^{\alpha}(1-t)^{\alpha} + \theta_2^{\alpha}t^{\alpha})^{\frac{1}{\alpha}} + (\theta_1 - \theta_2)t + 1 - \theta_1\]

References:

Bivariate extreme value theory: models and estimation. Biometrika, 1988.

AsymMixedCopula

TwinCopulas.AsymMixedCopula — Type

AsymMixedCopula{P}

Fields:

θ::Vector - parameters (size 2)

Constructor

AsymMixedCopula(θ)

The Asymmetric bivariate Mixed copula is parameterized by two parameters $\theta_{i}, i=1,2$ that must meet the following conditions:

θ₁ ≥ 0

θ₁+θ₂ ≤ 1

θ₁+2θ₂ ≤ 1

θ₁+3θ₂ ≥ 0

It is an Extreme value copula with Pickands dependence function:

\[A(t) = \theta_{2}t^3 + \theta_{1}t^2-(\theta_1+\theta_2)t+1\]

It has a few special cases:

When θ₁ = θ₂ = 0, it is the Independent Copula

When θ₂ = 0, it is the Mixed Copula

References:

Bivariate extreme value theory: models and estimation. Biometrika, 1988.

BC2Copula

TwinCopulas.BC2Copula — Type

BC2Copula{P}

Fields:

- θ1::Real - parameter
- θ1::Real - parameter

Constructor

BC2Copula(θ1, θ2)

The bivariate BC₂ copula is parameterized by two parameters $\theta_{i} \in [0,1], i=1,2$. It is an Extreme value copula with Pickands dependence function:

\[A(t) = \max\{\theta_1 t, \theta_2(1-t) \} + \max\{(1-\theta_1)t, (1-\theta_2)(1-t)\}\]

References:

Bivariate extreme-value copulas with discrete Pickands dependence measure. Springer, 2011.

CuadrasAugeCopula

TwinCopulas.CuadrasAugeCopula — Type

CuadrasAugeCopula{P}

Fields:

- α::Real - parameter

Constructor

CuadrasAugeCopula(α)

The bivariate Cuadras Auge copula is parameterized by $\alpha \in [0,1]$. It is an Extreme value copula with Pickands dependence function:

\[A(t) = \max\{t, 1-t \} + (1-\theta)\max\{t, 1-t\}\]

References:

Simulating copulas: stochastic models, sampling algorithms, and applications. 2017.

GalambosCopula

TwinCopulas.GalambosCopula — Type

GalambosCopula{P}

Fields:

- θ::Real - parameter

Constructor

GalambosCopula(θ)

The bivariate Galambos copula is parameterized by $\alpha \in [0,\infty)$. It is an Extreme value copula with Pickands dependence function:

\[A(t) = 1 - (t^{-\theta}+(1-t)^{-\theta})^{-\frac{1}{\theta}}\]

It has a few special cases:

When θ = 0, it is the Independent Copula

When θ = ∞, it is the M Copula (Upper Frechet-Hoeffding bound)

References:

Order statistics of samples from multivariate distributions. J. Amer. Statist Assoc. 1975.

HuslerReissCopula

TwinCopulas.HuslerReissCopula — Type

HuslerReissCopula{P}

Fields:

- θ::Real - parameter

Constructor

HuslerReissCopula(θ)

The bivariate Husler-Reiss copula is parameterized by $\theta \in [0,\infty)$. It is an Extreme value copula with Pickands dependence function:

\[A(t) = t\Phi(\theta^{-1}+\frac{1}{2}\theta\log(\frac{t}{1-t})) +(1-t)\Phi(\theta^{-1}+\frac{1}{2}\theta\log(\frac{1-t}{t}))\]

Where $\Phi$is the cumulative distribution function (CDF) of the standard normal distribution.

It has a few special cases:

When θ = 0, it is the Independent Copula

When θ = ∞, it is the M Copula (Upper Frechet-Hoeffding bound)

References:

Maxima of normal random vectors: between independence and complete dependence. Statist. Probab. 1989.

LogCopula

TwinCopulas.LogCopula — Type

LogCopula{P}

Fields:

- θ::Real - parameter

Constructor

LogCopula(θ)

The bivariate Logistic copula (or Gumbel Copula) is parameterized by $\theta \in [1,\infty)$. It is an Extreme value copula with Pickands dependence function:

\[A(t) = (t^{\theta}+(1-t)^{\theta})^{\frac{1}{\theta}}\]

It has a few special cases:

When θ = 1, it is the IndependentCopula

When θ = ∞, is is the MCopula (Upper Frechet-Hoeffding bound)

References:

Bivariate extreme value theory: models and estimation. Biometrika, 1988.

MixedCopula

TwinCopulas.MixedCopula — Type

MixedCopula{P}

Fields:

- θ::Real - parameter

Constructor

MixedCopula(θ)

The bivariate Mixed copula is parameterized by $\alpha \in [0,1]$. It is an Extreme value copula with Pickands dependence function:

\[A(t) = \theta t^2 - \theta t + 1\]

It has a few special cases:

When θ = 0, it is the IndependentCopula

References:

Bivariate extreme value theory: models and estimation. Biometrika, 1988.

MOCopula

TwinCopulas.MOCopula — Type

MOCopula{P}

Fields:

- λ1::Real - parameter
- λ2::Real - parameter
- λ12::Real - parameter

Constructor

MOCopula(θ)

The bivariate Marshall-Olkin copula is parameterized by $\lambda_i \in [0,\infty), i = 1, 2, \{1,2\}$. It is an Extreme value copula with Pickands dependence function:

\[A(t) = \frac{\lambda_1 (1-t)}{\lambda_1 + \lambda_{1,2}} + \frac{\lambda_2 t}{\lambda_2 + \lambda_{1,2}} + \lambda_{1,2}\max\left \{\frac{1-t}{\lambda_1 + \lambda_{1,2}}, \frac{t}{\lambda_2 + \lambda_{1,2}} \right \} \]

References:

Simulating copulas: stochastic models, sampling algorithms, and applications. 2017.

tEVCopula

TwinCopulas.tEVCopula — Type

tEVCopula{P}

Fields: - ν::Real - paremeter - θ::Real - Parameter

Constructor

tEVCopula(ν, θ)

The bivariate extreme t copula is parameterized by $\nu \in [0,\infty)$ and \theta \in (-1,1]. It is an Extreme value copula with Pickands dependence function:

\[A(x) = xt_{\nu+1}(Z_x) +(1-x)t_{\nu+1}(Z_{1-x})\]

Where $t_{\nu + 1}$is the cumulative distribution function (CDF) of the standard t distribution with \nu + 1 degrees of freedom and

\[Z_x = \frac{(1+\nu)^{1/2}{\sqrt{1-\theta^2}}\left [ \left (\frac{x}{1-x} \right )^{1/\nu} - \theta \right ]\]

It has a few special cases:

When θ = 0, it is the Independent Copula

When θ = ∞, it is the M Copula (Upper Frechet-Hoeffding bound)

References:

Extreme value properties of multivariate t copulas. Springer. 2008.