Other Copulas

RafteryCopula

TwinCopulas.RafteryCopula — Type

RafteryCopula{P}

Fields:

- θ::Real - parameter

Constructor

RafteryCopula(θ)

The bivariate Raftery Copula is parameterized by $\theta \in [0,1]$. It is constructed as:

\[C(u_1, u_2) = \min\{u_1,u_2\} +\frac{1-\theta}{1+\theta}(u_1u_2)^{1/(1-\theta)}\left \{1-(\max\{u_1,u_2\})^{-(1+\theta)/(1-\theta)} \right \}\]

It has a few special cases:

When θ = 0, it is the Independent Copula

When θ = 1, it is the MCopula (Upper Frechet-Hoeffding bound)

References:

Joe, Harry, Multivariate Models and Multivariate Dependence Concepts, Chapman & Hall. 1997.

B11Copula

TwinCopulas.B11Copula — Type

B11Copula{P}

Fields:

- θ::Real - parameter

Constructor

B11Copula(θ)

The bivariate B11 copula is parameterized by $\theta \in [0,1]$. It is constructed as:

\[C(u_1, u_2) = \theta \min\{u_1,u_2\} + (1-\theta)u_1u_2\]

It has a few special cases:

When θ = 0, it is the IndependentCopula

When θ = 1, is is the MCopula (Upper Frechet-Hoeffding bound)

References:

[Joe1997] Joe, Harry, Multivariate Models and Multivariate Dependence Concepts, Chapman & Hall. 1997.

FrechetCopula

TwinCopulas.FrechetCopula — Type

FrechetCopula{P}

Fields:

- θ1::Real - parameter
- θ2::Real - parameter

Constructor

FrechetCopula(θ1, θ2)

The bivariate Fréchet copula is parameterized by $\theta_{i} \in [0,1], i = 1,2$, such that $\theta_1 + \theta_2 \leq 1 \$. It is constructed as:

\[C(u_1, u_2) = \theta_1 \min\{u_1, u_2\} + (1- \theta_1 - \theta_2)u_1u_2 + \theta_2 \max\{u_1 + u_2 - 1, 0\}\]

It has a few special cases:

When θ1 = θ2 = 0, it is the Independent Copula

When θ1 = 1, it is the MCopula (Upper Frechet-Hoeffding bound)

When θ2 = 1, is is the WCopula (Lower Frechet-Hoeffding bound)

References:

Nelsen, Roger B. An introduction to copulas. Springer, 2006.

IndependentCopula

TwinCopulas.IndependentCopula — Type

IndependentCopula

Constructor

IndependentCopula()

The bivariate Independent Copula is the simplest copula, that has the form :

\[C(u_1,u_2) = u_1u_2\]

References:

Nelsen, Roger B. An introduction to copulas. Springer, 2006.

MardiaCopula

TwinCopulas.MardiaCopula — Type

MardiaCopula{P}

Fields:

- θ::Real - parameter

Constructor

MardiaCopula(θ)

The bivariate Mardia copula is parameterized by $\theta \in [-1,1]$. It is constructed as:

\[C(u_1, u_2) = \frac{\theta^2(1+\theta)}{2}\min\{u_1,u_2\} + (1-\theta^2)u_1u_2 + \frac{\theta^2(1-\theta)}{2}\max\{u_1+u_2-1,0\}\]

It has a few special cases:

When θ = 0, it is the Independent Copula

When θ = 1, it is the MCopula (Upper Frechet-Hoeffding bound)

When θ = -1, is is the WCopula (Lower Frechet-Hoeffding bound)

References:

Nelsen, Roger B. An introduction to copulas. Springer, 2006.

MaresiasCopula

TwinCopulas.MaresiasCopula — Type

MaresiasCopula{P}

Fields:

- G::Function

Constructor

MaresiasCopula(G)

The bivariate Maresias copula is parameterized by functions $G, H: [0,1] \to [0,1]$, such that $G(u) = 2u - H(u), u \in [0,1]$. It is constructed as:

\[C(u_1, u_2) = \frac{1}{2}\left ( G(u_1)G(u_2) + H(u_1)H(u_2) \right )\]

References:

Simulating copulas: stochastic models, sampling algorithms, and applications. 2017.

ArchimaxCopula

TwinCopulas.ArchimaxCopula — Type

ArchimaxCopula{A, E}

Fields:

-Archimedean::A - Archimedean Copula
-Extreme::E - Extreme Value Copula

Constructor

ArchimaxCopula(Archimedean, Extreme)

The bivariate Archimax Copula is parameterized by an Archimedean Copula and Extreme Value Copula. It is constructed as follows:

\[C_{\ell, \varphi}(u_1, u_2) = \varphi(\ell(\varphi^{-1}(u_1),\varphi^{-1}(u_2)))\]

where $\varphi$ is the generator of the Archimedean Copula and $\varphi^{-1}$ is the inverse and $\ell$ is the stable tail dependece function of Extreme value Copula

For more details see

References:

Charpentier et al, Multivariate Archimax Copulas, Journal of Multivariate analysis. 2014.

EmpiricalCopula

TwinCopulas.EmpiricalCopula — Type

EmpiricalCopula{M}

Fields:

-data::M - 2xn matrix observations

Constructor

EmpiricalCopula(data)

Let $(x_k, y_k), k = 1,2, \ldots, n$ denote a sample of size $n$ from continuous bivariate distribution. The Empirical Copula is the function

\[C_n(\frac{i}{n},\frac{j}{n}) = \frac{\text{number of pairs (x,y) in the sample with} x \leq x_{(i)}, y \leq y_{(j)}}{n}\]

For more details see

References:

Nelsen, Roger B. An introduction to copulas. Springer, 2006.

PlackettCopula

TwinCopulas.PlackettCopula — Type

PlackettCopula{P}

Fields: - θ::Real - parameter

Constructor

PlackettCopula(θ)

The bivariate Plackett Copula parameterized by $\theta > 0$ The Plackett. It is constructed as:

\[C(u_1,u_2) = \frac{1}{2}\eta^{-1}(1+\eta(u_1+u_2)-((1+\eta(u_1+u_2))^2 - 4\theta \eta u_1 u_2)^{1/2}) \]

where $\eta)=\theta-1.$

It has a few special cases:

When θ = ∞, is is the MCopula (Upper Frechet-Hoeffding bound)

When θ = 1, it is the IndependentCopula

When θ = 0, is is the WCopula (Lower Frechet-Hoeffding bound)

References:

Joe, H. (2014). Dependence modeling with copulas. CRC press, Page.164

Johnson, Mark E. Multivariate statistical simulation: A guide to selecting and generating continuous multivariate distributions. Vol. 192. John Wiley & Sons, 1987. Page 193.

Nelsen, Roger B. An introduction to copulas. Springer, 2006. Exercise 3.38.

SurvivalCopula

TwinCopulas.SurvivalCopula — Type

SurvivalCopula{P}

Fields:

- C::bicopula

Constructor

SurvivalCopula(C)

The bivariate Survival copula is defined as

\[\overbar{C}(u_1,u_2) = C(u_1, u_2) + u_1 + u_2 - 1\]

Where $C(u_1, u_2)$ is a copula as we know them.

References:

Nelsen, Roger B. An introduction to copulas. Springer, 2006.

MCopula

TwinCopulas.MCopula — Type

MCopula

Constructor

MCopula()

The Upper Frechet-Hoeffding bound is defined as

\[M(u_1,u_2) = \min\{u_1,u_2\}\]

References:

Nelsen, Roger B. An introduction to copulas. Springer, 2006.

MorgensternCopula

TwinCopulas.MorgensternCopula — Type

MorgensternCopula{P}

Fields:

- θ::Real - parameter

Constructor

MorgensternCopula(θ)

The bivariate Morgenstern copula or FGM Copula is parameterized by $\theta \in [-1,1]$. It is constructed as:

\[C(u_1, u_2) = u_1u_2(1+\theta(1-u_1)(1-u_2))\]

It has a few special cases:

When θ = 0, it is the Independent Copula

References:

Joe, Harry, Multivariate Models and Multivariate Dependence Concepts, Chapman & Hall. 1997.